ajustando arquivos da view e criando rotas

+++ controllers/task_controller.py
@@ -1,43 +1,41 @@
 from flask import render_template, request, redirect, url_for
 from models.task import Task, db
+from models.user import User
 
 class TaskController:
     @staticmethod
-    def index():
+    def list_tasks():
         tasks = Task.query.all()
-        return render_template('index.html', tasks=tasks)
+        return render_template('tasks.html', tasks=tasks)
 
     @staticmethod
-    def create():
+    def create_task():
         if request.method == 'POST':
             title = request.form['title']
             description = request.form['description']
             status = request.form['status']
+            user_id = request.form['user_id']
 
-            new_task = Task(title=title, description=description, status=status)
+            new_task = Task(title=title, description=description, status=status, user_id=user_id)
             db.session.add(new_task)
             db.session.commit()
 
-            return redirect(url_for('index'))
-
-        return render_template('create.html')
+            return redirect(url_for('list_tasks'))
+        users = User.query.all()
+        return render_template('create_task.html', users=users)
     
     @staticmethod
-    def update(task_id):
+    def update_task_status(task_id):
         task = Task.query.get(task_id)
-        if request.method == 'POST':
-            task.title = request.form['title']
-            task.description = request.form['description']
-            task.status = request.form['status']
-
+        if task:
+            task.status = 'Concluído' if task.status != 'Concluído' else 'Pendente'
             db.session.commit()
-            return redirect(url_for('index'))
-
-        return render_template('update.html', task=task)
+        return redirect(url_for('list_tasks'))
     
     @staticmethod
-    def delete(task_id):
+    def delete_task(task_id):
         task = Task.query.get(task_id)
-        db.session.delete(task)
-        db.session.commit()
-        return redirect(url_for('index'))
+        if task:
+            db.session.delete(task)
+            db.session.commit()
+        return redirect(url_for('list_tasks'))
